+++ retdec/tools/decompiler.py
@@ -276,6 +276,12 @@ def parse_args(argv):
         help='print fewer information during the decompilation'
     )
     parser.add_argument(
+        '-c', '--compiler',
+        dest='comp_compiler',
+        choices=['gcc', 'clang'],
+        help='compiler to be used when compiling'
+    )
+    parser.add_argument(
         '-f', '--file-format',
         dest='file_format',
         choices=['elf', 'pe'],
@@ -383,6 +389,7 @@ def main(argv=None):
     add_decompilation_param_when_given(args, params, 'target_language')
     add_decompilation_param_when_given(args, params, 'architecture')
     add_decompilation_param_when_given(args, params, 'file_format')
+    add_decompilation_param_when_given(args, params, 'comp_compiler')
     add_decompilation_param_when_given(args, params, 'generate_archive')
     decompilation = decompiler.start_decompilation(**params)